+++ src/app/_component/LottieLayout.tsx
