+++ src/app/_hoc/WithAuth.tsx
@@ -1,13 +1,13 @@
 "use client";
 
-import LottieLayout from "@components/layout/LottieLayout";
+import LottieLayout from "@app/_component/LottieLayout";
 import useLogin from "@hooks/login";
 import { Player } from "@lottiefiles/react-lottie-player";
 import { TUser } from "@t/user.type";
 import { useRouter } from "next/navigation";
 import { useEffect, useState } from "react";
 
-import Lottie from "../../public/lottie/loading.json";
+import Lottie from "../../../public/lottie/loading.json";
 
 export default function WithAuth(WrappedComponent) {
     return function ChildComponent(props) {
