+++ src/app/rooms/[roomId]/_component/DialogList.tsx
@@ -1,0 +1,1 @@
+export default function DialogList() {}
