+++ src/app/rooms/[roomId]/_component/Textator.tsx
